% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data-helper.R
\name{api_url}
\alias{api_url}
\title{Getting data from the IMD api}
\usage{
api_url()
}
\value{
dataset
}
\description{
Only required for IMD as postcodes is routed through the {NHSRpostcodetools}
package
}
