% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode-helpers.R
\name{check_postcodes}
\alias{check_postcodes}
\title{Use {NHSRpostcodetools} to validate postcodes}
\usage{
check_postcodes(data)
}
\arguments{
\item{data}{dataframe or vector}
}
\value{
data frame with data from the original along with postcode
information
}
\description{
Use {NHSRpostcodetools} to validate postcodes
}
