% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode-helpers.R
\name{get_api}
\alias{get_api}
\title{Send a GET request to the API}
\usage{
get_api(query_params)
}
\arguments{
\item{query_params}{uses the output from the function \code{set_country_params()}}
}
\value{
list
}
\description{
Send a GET request to the API
}
