% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_count}
\alias{get_count}
\title{Return the vector of country count of LSOAs}
\usage{
get_count(response_list)
}
\arguments{
\item{response_list}{}
}
\value{
vector
}
\description{
Return the vector of country count of LSOAs
}
