% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-counts.R
\name{get_country_count}
\alias{get_country_count}
\title{Get the country count}
\usage{
get_country_count()
}
\value{
vector
}
\description{
relies on function \code{set_country_params} that needs to be split out to
different nations
}
