% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_data}
\alias{get_data}
\title{Create a data frame}
\usage{
get_data(response_list)
}
\arguments{
\item{response_list}{uses output from the function \code{parse_json()}}
}
\value{
data frame
}
\description{
Create a data frame
}
