% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imd.R
\name{get_imd}
\alias{get_imd}
\title{Function that validates postcodes and links to postcode dataset which
includes IMD, returning only the score and generated deciles and quintiles.}
\usage{
get_imd(x, var = "postcode", fix_invalid = TRUE)
}
\arguments{
\item{x}{A data frame with a column of postcodes, or a vector
of postcodes.}

\item{var}{String or symbol. The name of the variable in the data frame that
comprises the postcodes to be submitted.
Should be acceptable as a symbol or as a standard string.}

\item{fix_invalid}{Boolean, default \code{TRUE}. Whether to try to fix any
postcodes that are not found (potentially because they are terminated codes,
or contain typos).}
}
\description{
Function that validates postcodes and links to postcode dataset which
includes IMD, returning only the score and generated deciles and quintiles.
}
