% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imd.R
\name{get_imd_api}
\alias{get_imd_api}
\title{Function that links to IMD dataset. Currently returns only a maximum of
2000 rows and should be just over 4000}
\usage{
get_imd_api(x)
}
\arguments{
\item{x}{A data frame with a column of postcodes, or a vector
of postcodes.}
}
\description{
Function that links to IMD dataset. Currently returns only a maximum of
2000 rows and should be just over 4000
}
