% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode-helper.R
\name{is_lsoa}
\alias{is_lsoa}
\title{Function to check if a string is an LSOA code}
\usage{
is_lsoa(LSOA)
}
\arguments{
\item{LSOA}{string}
}
\value{
Logical
}
\description{
Function to check if a string is an LSOA code
}
