% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode-helper.R
\name{is_postcode}
\alias{is_postcode}
\title{Function to check if a string is a valid postcode regardless of its
formatting.}
\usage{
is_postcode(data)
}
\arguments{
\item{data}{string}
}
\value{
Logical
}
\description{
With or without one space is permitted.
Does not work with data where two spaces are included with 6 character
postcodes.
From the MIT licenced package {Rbduk}
https://github.com/sama-ds/Rbduk/blob/main/R/is_postcode.R
}
\examples{
is_postcode("SW1a2nP")
is_postcode("SW1a 2nP")
}
