% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode-helpers.R
\name{parse_json}
\alias{parse_json}
\title{Parse the JSON response}
\usage{
parse_json(json)
}
\arguments{
\item{json}{uses output from function \code{get_api()}}
}
\description{
Parse the JSON response
}
