% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imd-helpers.R
\name{parse_json_imd}
\alias{parse_json_imd}
\title{Parse the imd JSON response}
\usage{
parse_json_imd(json)
}
\arguments{
\item{json}{uses output from function \code{get_api_imd()}}
}
\description{
Parse the imd JSON response
}
