% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{pull_table_data}
\alias{pull_table_data}
\title{pull actual data out from API JSON response}
\usage{
pull_table_data(resp)
}
\arguments{
\item{respused}{in function \code{\link{get_data}}}
}
\value{
function
}
\description{
pull actual data out from API JSON response
}
