% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{retrieve_data}
\alias{retrieve_data}
\title{use batched IDs to retrieve table data}
\usage{
retrieve_data(req, ids_vec)
}
\arguments{
\item{req}{used in function \code{\link{get_data}}}

\item{ids_vec}{used in function \code{\link{get_data}}}
}
\value{
function
}
\description{
use batched IDs to retrieve table data
}
