% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode-helpers.R
\name{set_country_params}
\alias{set_country_params}
\title{Country parameters to return the total number of LSOAs}
\usage{
set_country_params()
}
\value{
list of parameters required, currently England only
}
\description{
Deciles and Quintiles are not available from source and need to be #
calculated from the total number of LSOAs for the nation.

TODO expand to nations depending on starting letter of LSOA
England = E
Wales = W
Scotland = S
Northern Ireland = (9

https://statistics.ukdataservice.ac.uk/dataset/2011-census-geography-boundaries-lower-layer-super-output-areas-and-data-zones
}
