% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_params}
\alias{set_params}
\title{Connect to services1.arcgis API for data}
\usage{
set_params(data, type = c("postcode", "lsoa"))
}
\arguments{
\item{data}{data frame output from {NHSRpostcodetools}}

\item{type}{select which parameters to change}
}
\value{
list of parameters required, currently PCDS
}
\description{
Connect to services1.arcgis API for data
}
