% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{api_url}
\alias{api_url}
\title{Getting data from the api_urls}
\usage{
api_url(url_type = c("postcode", "imd", "lsoa"))
}
\arguments{
\item{url_type}{String defaults to \code{postcode} which connects to
Online_ONS_Postcode_Directory_Live and \code{imd} connects to
Indices_of_Multiple_Deprivation_(IMD)_2019}
}
\value{
dataset
}
\description{
Getting data from the api_urls
}
