% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Query information to restrict data returned}
\usage{
get_data(
  data,
  url_type = c("postcode", "imd"),
  fix_invalid = TRUE,
  column = "default"
)
}
\arguments{
\item{data}{dataframe or vector.The data that will connect to either the
postcode API or imd API.}

\item{url_type}{String defaults to \code{postcode} which returns information from
Online_ONS_Postcode_Directory_Live to return Postcode information via the
{NHSRpostcodetools} package. \code{imd} connects to
Indices_of_Multiple_Deprivation_(IMD)_2019 to return IMD information.}

\item{fix_invalid}{Boolean, default \code{TRUE}. Whether to try to fix any
postcodes that are not found (potentially because they are terminated codes,
or contain typos).}

\item{column}{String. Default would mean the automatic connection of a
column called \code{postcode} if postcode data is being expected or \code{lsoa11} if
imd data as requested via the parameter \code{url_type = "imd"}.}
}
\description{
This function queries the API based on the information required and returns
ids as these have no restriction applied to them.
The IMD API restricts to 2k records for example.

Postcode and LSOA parameters require data because bringing all the data
from the Online_ONS_Postcode_Directory_Live will take too long and is often
unnecessary.
}
