% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Query information to restrict data returned}
\usage{
get_data(data, url_type = c("postcode", "imd", "lsoa"))
}
\arguments{
\item{data}{String in "postcode" this is checked using the
{NHSRpostcodetools} so selects the postcode column from a data frame or can
handle a vector.
This functionality needs to be extended to "lsoa" codes which currently only
allow a vector.
"imd" brings all data back and only relates to the English IMD.}

\item{url_type}{String "postcode" and "lsoa" connect to
Online_ONS_Postcode_Directory_Live and imd connects to
Indices_of_Multiple_Deprivation_(IMD)_2019}
}
\value{
list Ids only as these are faster to batch
}
\description{
This function queries the API based on the information required and returns
ids as these have no restriction applied to them.
The IMD API restricts to 2k records for example.

Postcode and LSOA parameters require data because bringing all the data
from the Online_ONS_Postcode_Directory_Live will take too long and is often
unnecessary.
}
