% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imd.R
\name{get_imd}
\alias{get_imd}
\title{Function that validates postcodes and links to IMD datasets, returning the
score and generated deciles and quintiles.}
\usage{
get_imd(.data, var = "postcode", fix_invalid = TRUE)
}
\arguments{
\item{.data}{A data frame with a column of postcodes, or a vector
of postcodes.}

\item{var}{String or symbol. The name of the variable in the data frame that
comprises the postcodes to be submitted.
Should be acceptable as a symbol or as a standard string.}

\item{fix_invalid}{Boolean, default \code{TRUE}. Whether to try to fix any
postcodes that are not found (potentially because they are terminated codes,
or contain typos).}

\item{country}{select which country in the nations the IMD scores relate to.
Note that datasets cannot be mixed.
Deciles are generated from the data provided from the API and so are based
on the totals.}
}
\description{
Function that validates postcodes and links to IMD datasets, returning the
score and generated deciles and quintiles.
}
\examples{
postcodes <- c("HD1 2UT", "HD1 2UU", "HD1 2UV")
test_df1 <- dplyr::tibble(
  place = paste0("place_", 1:3),
  postcode = postcodes
)
postcode_data_join(test_df1, fix_invalid = TRUE)
}
