% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_imd}
\alias{get_imd}
\title{Lower layer Super Output Area (LSOA) population estimates for England}
\format{
A data frame with 6325046 rows and 10 variables:
\describe{
\item{lsoa_year}{Numeric, the year of the UK Census that the LSOA are
linked to}
\item{lsoa_code}{Character, Lower layer Super Output Area (LSOA) code}
\item{lsoa_name}{Character, Lower layer Super Output Area (LSOA) name}
\item{la_year}{Numeric, year indicating when the Local Authority (LA)
boundaries were last defined/updated}
\item{la_code}{Character, Local Authority (LA) code}
\item{la_name}{Character, Local Authority (LA) name}
\item{age}{Numeric, age groups from 0 to 90+, not that the age category 90
includes estimates for the ages of 90+}
\item{gender}{Character, gender (f = female; m = male)}
\item{est_year}{Numeric, year of ONS estimate}
\item{n}{Number of the estimated population}
}
}
\source{
\url{https://services1.arcgis.com/ESMARspQHYMw9BZ9/ArcGIS/rest/services/Online_ONS_Postcode_Directory_Live/FeatureServer/0/query?}
}
\usage{
get_imd(.data, var = "postcode", fix_invalid = TRUE)
}
\description{
A dataset containing Lower layer Super Output Area (LSOA) population
estimates for England
}
\keyword{dataset}
