% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imd.R
\name{get_lsoa}
\alias{get_lsoa}
\title{Function that gets IMD deciles and quintiles for LSOA codes.}
\usage{
get_lsoa(.data, return = c("all", "random", "first"))
}
\arguments{
\item{.data}{A data frame with a column of lsoas, or a vector
of lsoas.}

\item{return}{data frame of all postcodes in the LSOA area as default,
one random from a decile or the first from a decile for example postcodes.}
}
\description{
Function that gets IMD deciles and quintiles for LSOA codes.
}
